requirements.txt
# Core API framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Neo4j database driver
neo4j==5.15.0

# LLM Integration
anthropic>=0.45.2

# Additional FastAPI dependencies
pydantic==2.5.0
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0
